Extract gitpanel/ sub-package from internal/tui
Move gitstatus.go and fileexplorer.go to internal/tui/gitpanel/.
Add Focused() accessor on FilePanel for cross-package test access.

Reviews and settings remain in root tui/ for now — they have deep
App coupling that requires interface extraction in a follow-up.

Phase 5 (partial) of the tui package decomposition.
