+++ internal/tui/gitpanel/fileexplorer_test.go
@@ -1,4 +1,4 @@
-package tui
+package gitpanel
 
 import (
 	"testing"
