+++ internal/tui/gitpanel/gitstatus.go
@@ -1,4 +1,4 @@
-package tui
+package gitpanel
 
 import (
 	"strings"
